requirements.txt
alembic==0.8.9
aniso8601==6.0.0
bcrypt==3.1.2
cffi==1.9.1
click==6.6
coverage==4.2
Flask==0.12
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-Migrate==2.0.2
Flask-RESTful==0.3.7
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.6.1
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.6
MarkupSafe==0.23
mysqlclient==1.4.2.post1
pycparser==2.17
PyJWT==1.4.2
python-dotenv==0.10.1
python-editor==1.0.3
pytz==2018.9
six==1.10.0
SQLAlchemy==1.1.4
Werkzeug==0.11.13
